+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [0.1.1]
 - added cr reactive planner integration via method for global trajectory
+- support of commonroad-route-planner 2025.0.1 api changes
 - added fast api